+++ ledger/tests/test_decarators.py
@@ -55,10 +55,17 @@ class TestDecorators(TestCase):
 
     def test_log_timing(self):
         # given
+        # Alliance Auth
+        from allianceauth.services.hooks import get_extension_logger
+
+        # Alliance Auth (External Libs)
+        from app_utils.logging import LoggerAddTag
+
         # AA Ledger
+        from ledger import __title__
         from ledger.decorators import log_timing
 
-        logger = logging.getLogger(__name__)
+        logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
         @log_timing(logger)
         def trigger_log_timing():