+++ ledger/views/character/add_char.py
@@ -2,9 +2,6 @@
 Character Audit
 """
 
-# Standard Library
-import logging
-
 # Django
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
@@ -13,13 +10,17 @@ from django.utils.translation import gettext_lazy as trans
 
 # Alliance Auth
 from allianceauth.eveonline.models import EveCharacter
+from allianceauth.services.hooks import get_extension_logger
 from esi.decorators import token_required
 
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
 # AA Ledger
-from ledger import tasks
+from ledger import __title__, tasks
 from ledger.models.characteraudit import CharacterAudit
 
-logger = logging.getLogger(__name__)
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 @login_required